server/requirements.txt
black==19.3b0
django==3.0.4
psycopg2-binary==2.8.4
Pillow==7.0.0
django-cors-headers==3.2.1
django-jwt-auth==0.0.2
daphne==2.4.1
PyJWT==1.7.1
gevent==1.4.0
graphene==2.1.8
graphql-ws-django==0.3.4
graphene-django==2.9.0
graphql-core<3
graphql-relay<3
gunicorn==20.0.4
django-extensions==2.2.8
ipdb==0.13.2
serious-django-graphene==0.5.1
django-graphql-jwt==0.3.0
celery==4.4.2
redis==3.4.1
flower==0.9.3
Rx<3
six==1.14.0
channels==2.4.0
channels_redis~=2.3.3
tornado>=4.2.0,<6.0.0